+++ packages/terra-card/src/CardBody.jsx
@@ -19,22 +19,29 @@ const propTypes = {
    * Provides themeable padding horizontal
    */
   hasPaddingHorizontal: PropTypes.bool,
+  /**
+  * Sets the content of the card to be centered
+  */
+  isContentCentered: PropTypes.bool,
 };
 
 const defaultProps = {
   hasPaddingHorizontal: true,
   hasPaddingVertical: true,
+  isContentCentered: false,
 };
 
 const CardBody = ({
   children,
   hasPaddingVertical,
   hasPaddingHorizontal,
+  isContentCentered,
   ...customProps
 }) => {
   const cardBodyClasses = cx([
     { 'vertical-padding': hasPaddingVertical },
     { 'horizontal-padding': hasPaddingHorizontal },
+    { 'center-content': isContentCentered },
     customProps.className,
   ]);
   return <div {...customProps} className={cardBodyClasses}>{children}</div>;