+++ packages/terra-card/tests/nightwatch/CardContentCenter.jsx
@@ -1,0 +1,11 @@
+import React from 'react';
+
+import Card from '../../lib/Card';
+import CardBody from '../../lib/CardBody';
+
+export default () =>
+  <Card>
+    <CardBody isContentCentered>
+      Centered content
+    </CardBody>
+  </Card>;